resolver: inline ttl_config_deserialize module
